Fixed app so not all game logic is in App.tsx
